chore: rename api-utils as api-utils-client
